pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "michael-doors-bot"
version = "1.0.0"
description = "WhatsApp AI bot for Michael Doors business"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "anthropic>=0.28.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
start = "uvicorn src.michael_doors_bot.main:app --host 0.0.0.0 --port ${PORT:-3000}"
